задача № 12334

+++ resources/views/site/catalog.blade.php
@@ -119,19 +119,19 @@
                         <tr>
                             <td>{{trans('catalog.Economy (7 days)')}}</td>
                             <td>@if(App::getLocale() == 'en')
-                                    0.07 @else {{0.07 * CurrencyRate::getOneRecord()}} @endif</td>
+                                    0.07 @else {{ (double) 0.07 * CurrencyRate::getOneRecord()}} @endif</td>
                         </tr>
 
 
                         <tr>
                             <td>{{trans('catalog.Standard (4 days)')}}</td>
                             <td>@if(App::getLocale() == 'en')
-                                    0.08 @else {{0.08 * CurrencyRate::getOneRecord()}} @endif</td>
+                                    0.08 @else {{ (double) 0.08 * CurrencyRate::getOneRecord()}} @endif</td>
                         </tr>
                         <tr>
                             <td>{{trans('catalog.Express (2 days)')}}</td>
                             <td>@if(App::getLocale() == 'en')
-                                    0.09 @else {{0.09 * CurrencyRate::getOneRecord()}} @endif</td>
+                                    0.09 @else {{ (double) 0.09 * CurrencyRate::getOneRecord()}} @endif</td>
                         </tr>
                         </tbody>
                     </table>
@@ -153,19 +153,19 @@
                         <tr>
                             <td>{{trans('catalog.Economy (10 days)')}}</td>
                             <td> @if(App::getLocale() == 'en')
-                                    0.10 @else {{0.10 * CurrencyRate::getOneRecord()}} @endif</td>
+                                    0.10 @else {{ (double) 0.10 * CurrencyRate::getOneRecord()}} @endif</td>
                         </tr>
 
 
                         <tr>
                             <td>{{trans('catalog.Standard (7 days)')}}</td>
                             <td>@if(App::getLocale() == 'en')
-                                    0.11 @else {{0.11 * CurrencyRate::getOneRecord()}} @endif</td>
+                                    0.11 @else {{ (double) 0.11 * CurrencyRate::getOneRecord()}} @endif</td>
                         </tr>
                         <tr>
                             <td>{{trans('catalog.Express (4 days)')}}</td>
                             <td>@if(App::getLocale() == 'en')
-                                    0.12 @else {{0.12 * CurrencyRate::getOneRecord()}} @endif</td>
+                                    0.12 @else {{ (double) 0.12 * CurrencyRate::getOneRecord()}} @endif</td>
                         </tr>
                         </tbody>
                     </table>
@@ -190,7 +190,7 @@
                             <td>0-5000</td>
                             <td>2 {{trans('catalog.day')}}</td>
                             <td>@if(App::getLocale() == 'en')
-                                    110 @else {{110 * CurrencyRate::getOneRecord()}} @endif</td>
+                                    110 @else {{ (double) 110 * CurrencyRate::getOneRecord()}} @endif</td>
                         </tr>
 
 
@@ -199,14 +199,14 @@
                             <td>5000-10000</td>
                             <td>4 {{trans('catalog.day')}}</td>
                             <td>@if(App::getLocale() == 'en')
-                                    190 @else {{190 * CurrencyRate::getOneRecord()}} @endif</td>
+                                    190 @else {{ (double) 190 * CurrencyRate::getOneRecord()}} @endif</td>
                         </tr>
                         <tr>
 
                             <td>10000-15000</td>
                             <td>4-10 {{trans('catalog.day')}}</td>
                             <td>@if(App::getLocale() == 'en')
-                                    270 @else {{270 * CurrencyRate::getOneRecord()}} @endif</td>
+                                    270 @else {{ (double) 270 * CurrencyRate::getOneRecord()}} @endif</td>
                         </tr>
 
 
@@ -238,7 +238,8 @@
 
                             <td>{{trans('catalog.Formatting tables and graphs (one illustration of an unlimited number of tables and charts in it)')}}</td>
                             <td>{{trans('catalog.By agreement')}}</td>
-                            <td>@if(App::getLocale() == 'en') 80 @else {{80 * CurrencyRate::getOneRecord()}} @endif</td>
+                            <td>@if(App::getLocale() == 'en')
+                                    80 @else {{  (double)  80 * CurrencyRate::getOneRecord()}} @endif</td>
                         </tr>
